SGP Models

The sgp_models module contains the implementation of the Simplified General Perturbations (SGP) models for orbit propagation. These models are based on the work of Vallado et al. (2001). The SGP models are a set of simplified models for propagating the orbits of Earth-orbiting satellites. The models are based on the two-line element (TLE) format used by the US Department of Defense to distribute satellite orbit data. The SGP models are widely used in the satellite tracking community due to their simplicity and computational efficiency.

SatelliteDynamics.TLE — Type

Structure for storing a NORAD Two-Line Element

Atributes:

line1::String First line of Two-Line Element set including checksum

line2::String Second line of Two-Line Element set including checksum

epoch::Epoch Epoch of element set.

SatelliteDynamics.state — Function

Compute the satellite state at the given epoch as output from the SGP4 propagator.

Arguments:

tle::TLE Two-Line element object

epc::Epoch Epoch for SGP4 state

opscode::Char Operatitonal mode of propagators

Returns:

rv::AbstractArray{Float64, 1} Position and velocity as output by TLE propagator. Units [m m/s]

SatelliteDynamics.ecef — Function

Compute the satellite state in the Earth-fixed frame.

Arguments:

tle::TLE Two-Line element object

epc::Epoch Epoch for SGP4 state

Returns:

ecef::AbstractArray{Float64, 1} Position and velocity in ECEF frame. Units [m; m/s]

SatelliteDynamics.eci — Function

Compute the satellite state in the Earth-centered inertial.

Arguments:

tle::TLE Two-Line element object

epc::Epoch Epoch for SGP4 state

Returns:

eci::AbstractArray{Float64, 1} Position and velocity in ECI frame. Units [m; m/s]