Time

Time

The Time submodule contains common time transformations such as converting between different date representations or converting a specific instant in time between different time systems.

The module also defines the Epoch class which p

Most of the transformations are make backend calls to the SOFA C-library functions provide the package SOFA.jl

SatelliteDynamics.Time.caldate_to_mjd — Function.

Convert a Gregorian calendar date to the equivalent Modified Julian Date representation of that time instant.

Aguments:

year::Int Year

year::Int Month

year::Int Day

hour::Int Hour

minute::Int Minute

second::Real Seconds

nanoseconds::Real Microseconds

Returns:

mjd::Float64 Modified Julian Date of Epoch

source

SatelliteDynamics.Time.mjd_to_caldate — Function.

Convert a Modified Julian Date to the equivalent Gregorian calendar date representation of the same instant in time.

Aguments:

mjd::Real: Modified Julian Date of Epoch

Returns:

year::Int32: Year

year::Int32: Month

year::Int32: Day

hour::Int32: Hour

minute::Int32: Minute

second::Float64: Seconds

nanoseconds::Float64: nanosecondss

source

SatelliteDynamics.Time.caldate_to_jd — Function.

Convert a Gregorian calendar date to the equivalent Julian Date representation of that time instant.

Aguments:

year::Int: Year

year::Int: Month

year::Int: Day

hour::Int: Hour

minute::Int: Minute

second::Real: Seconds

nanoseconds::Real: nanosecondss

Returns:

mjd::Float64: Julian Date of Epoch

source

SatelliteDynamics.Time.jd_to_caldate — Function.

Convert a Julian Date to the equivalent Gregorian calendar date representation of the same instant in time.

Aguments:

mjd::Real: Julian Date of Epoch

Returns:

year::Int32: Year

year::Int32: Month

year::Int32: Day

hour::Int32: Hour

minute::Int32: Minute

second::Float64: Seconds

microsecond::Float64: Microseconds

source

SatelliteDynamics.Time.elapsed_from_epoch — Function.

Compute the number of elapsed seconds since a given Epoch from the day number. Can be used to compute the elapsed time since a given Julian or Modified Julian Date.

Arguments:

day_number::Real: Day number, can contain fractional days. Asummes all days are a uniform 86400.0 seconds in length.

day_epoch::Real: Modified Julian Date of Epoch

Returns:

t::Float: Number of elapsed seconds between the Provided Modified Julian date and the epoch.

source

SatelliteDynamics.Time.days_from_elapsed — Function.

Computes the day number in a given time scale given the elapsed time since epoch and the epoch itself.

Assumes all days are counted using a uniform 86400.0 seconds over the time span.

Arguments:

t::Real: Elapsed seconds since the day_epoch.

day_epoch::Real: Day number of the epoch. Common values are SatelliteDynamics.Constants.MJD_ZERO (to get the Julian Day number) or SatelliteDynamics.Constants.MJD2000 (to get Modified Julian Days if reckoning time from January 1, 2000 0H)

Returns:

days::Float: Number of elapsed days in the time scale.

source