Constants

The constants submodule constains common mathematical, physical, astronomical constants.

SatelliteDynamics.AS2RAD — Constant

Constant to convert arcseconds to radians. Equal to 2pi/(360*3600). [rad/as]

SatelliteDynamics.RAD2AS — Constant

Constant to convert radians to arcseconds. Equal to 2pi/(360*3600) [as/ras]

SatelliteDynamics.C_LIGHT — Constant

Speed of light in vacuum. [m/s]

D. Vallado, Fundamentals of Astrodynamics and Applications (4th Ed.), 2010

SatelliteDynamics.AU — Constant

Astronomical Unit. Equal to the mean distance of the Earth from the sun. TDB-compatible value. [m]

P. Gérard and B. Luzum, IERS Technical Note 36, 2010

SatelliteDynamics.MJD_ZERO — Constant

Offset of Modified Julian Days representation with respect to Julian Days. For a time, t, MJD_ZERO is equal to:

MJD_ZERO = t_jd - t_mjd

Where tjd is the epoch represented in Julian Days, and tmjd is the epoch in Modified Julian Days.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.MJD2000 — Constant

Modified Julian Date of January 1, 2000 00:00:00. Value is independent of time scale.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.GPS_TAI — Constant

Offset of GPS time system with respect to TAI time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.TAI_GPS — Constant

Offset of TAI time system with respect to GPS time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.TT_TAI — Constant

Offset of TT time system with respect to TAI time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.TAI_TT — Constant

Offset of TAI time system with respect to TT time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.GPS_TT — Constant

Offset of GPS time system with respect to TT time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.TT_GPS — Constant

Offset of TT time system with respect to GPS time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.GPS_ZERO — Constant

Modified Julian Date of the start of the GPS time system in the GPS time system. This date was January 6, 1980 0H as reckond in the UTC time system.

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.R_EARTH — Constant

Earth's equatorial radius. [m]

GGM05s Gravity Model

SatelliteDynamics.WGS84_a — Constant

Earth's semi-major axis as defined by the WGS84 geodetic system. [m]

NIMA Technical Report TR8350.2

SatelliteDynamics.WGS84_f — Constant

Earth's ellipsoidal flattening. WGS84 Value.

NIMA Technical Report TR8350.2

SatelliteDynamics.GM_EARTH — Constant

Earth's Gravitational constant [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.e_EARTH — Constant

Earth's first eccentricity. WGS84 Value. [dimensionless]

NIMA Technical Report TR8350.2

SatelliteDynamics.J2_EARTH — Constant

Earth's first zonal harmonic. [dimensionless]

GGM05s Gravity Model.

SatelliteDynamics.OMEGA_EARTH — Constant

Earth axial rotation rate. [rad/s]

D. Vallado, Fundamentals of Astrodynamics and Applications (4th Ed.), p. 222, 2010

SatelliteDynamics.GM_SUN — Constant

Gravitational constant of the Sun. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.R_SUN — Constant

Nominal solar photospheric radius. [m]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.P_SUN — Constant

Nominal solar radiation pressure at 1 AU. [N/m^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.GM_MOON — Constant

Gravitational constant of the Moon. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.GM_MERCURY — Constant

Gravitational constant of the Mercury. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.GM_VENUS — Constant

Gravitational constant of the Venus. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.GM_MARS — Constant

Gravitational constant of the Mars. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.GM_JUPITER — Constant

Gravitational constant of the Jupiter. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.GM_SATURN — Constant

Gravitational constant of the Saturn. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.GM_URANUS — Constant

Gravitational constant of the Uranus. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.GM_NEPTUNE — Constant

Gravitational constant of the Neptune. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.

SatelliteDynamics.GM_PLUTO — Constant

Gravitational constant of the Pluto. [m^3/s^2]

O. Montenbruck, and E. Gill, Satellite Orbits: Models, Methods and Applications, 2012.