Coordinates

Coordinates

The coordiantes module provides functionatiy for transforming between various body-fixed reference frame coordinate systems. In particular, geodetic and geocentric transformations are included. Topocentric transformations are also included.

SatelliteDynamics.Coordinates.sGEOCtoECEF — Function.

Convert geocentric position to equivalent Earth-fixed position.

Arguments:

geoc::Array{<:Real, 1}: Geocentric coordinates (lon, lat, altitude) [rad] / [deg]

use_degrees:Bool: If true interpret input as being in degrees.

Returns:

ecef::Array{<:Real, 1}: Earth-fixed coordinates [m]

source

SatelliteDynamics.Coordinates.sECEFtoGEOC — Function.

Convert Earth-fixed position to geocentric location.

Arguments:

ecef::Array{<:Real, 1}: Earth-fixed coordinated [m]

use_degrees:Bool: If true returns result in units of degrees

Returns:

geoc: Geocentric coordinates (lon, lat, altitude) [rad] / [deg]

source

SatelliteDynamics.Coordinates.sGEODtoECEF — Function.

Convert geodetic position to equivalent Earth-fixed position.

Arguments:

geod::Array{<:Real, 1}: Geodetic coordinates (lon, lat, altitude) [rad] / [deg]

use_degrees:Bool: If true interpret input as being in degrees.

Returns:

ecef::Array{<:Real, 1}: Earth-fixed coordinates [m]

source

SatelliteDynamics.Coordinates.sECEFtoGEOD — Function.

Convert geodetic coordinaties to Earth-fixed position

Arguments:

ecef::Array{<:Real, 1}: Earth-fixed position [m]

use_degrees:Bool: If true returns result in units of degrees

Returns:

geod::Array{<:Real, 1}: Geocentric coordinates (lon, lat, altitude) [rad] / [deg]

source

SatelliteDynamics.Coordinates.rECEFtoENZ — Function.

Compute the rotation matrix from the Earth-fixed to the East-North-Up coorindate basis.

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Real, 2}: Topocentric rotation matrix

source

SatelliteDynamics.Coordinates.rENZtoECEF — Function.

Compute the rotation matrix from the Earth-fixed to the South-East-Zenith coorindate basis.

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

source

SatelliteDynamics.Coordinates.sECEFtoENZ — Function.

Compute the coordinates of an object in the topocentric frame of an Earth-fixed frame

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

ecef::Array{<:Real, 1}: Coordinates of the object in Earth-fixed station

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

source

SatelliteDynamics.Coordinates.sENZtoECEF — Function.

Compute the coordinates of an object in the topocentric frame of an Earth-fixed frame

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

sez::Array{<:Real, 1}: SEZ coordinates of the object

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

source

SatelliteDynamics.Coordinates.rECEFtoSEZ — Function.

Compute the rotation matrix from the Earth-fixed to the South-East-Zenith coorindate basis.

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

source

SatelliteDynamics.Coordinates.rSEZtoECEF — Function.

Compute the rotation matrix from the Earth-fixed to the South-East-Zenith coorindate basis.

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

source

SatelliteDynamics.Coordinates.sECEFtoSEZ — Function.

Compute the coordinates of an object in the topocentric frame of an Earth-fixed frame

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

ecef::Array{<:Real, 1}: Coordinates of the object in Earth-fixed station

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

source

SatelliteDynamics.Coordinates.sSEZtoECEF — Function.

Compute the coordinates of an object in the topocentric frame of an Earth-fixed frame

Arguments:

station_ecef::Array{<:Real, 1}: Earth-fixed cartesian station coordinates

sez::Array{<:Real, 1}: SEZ coordinates of the object

conversion::Bool: Conversion type to use. Can be "geocentric" or "geodetic"

Returns:

E::Array{Float64, 2}: Topocentric rotation matrix

source

SatelliteDynamics.Coordinates.sENZtoAZEL — Function.

Convert East-North-Zenith topocentric state to azimuth, elevation, and range.

Arguments:

x::Array{<:Real, 1}: East-North-Up state

use_degrees:Bool: If true returns result in units of degrees

Returns:

azel::Array{<:Real, 1}: Azimuth, elevation and range [rad; rad; m]

source

SatelliteDynamics.Coordinates.sSEZtoAZEL — Function.

Convert South-East-Zenith topocentric state to azimuth, elevation, and range.

Arguments:

x::Array{<:Real, 1}: South-East-Zenith state

use_degrees:Bool: If true returns result in units of degrees

Returns:

azel::Array{<:Real, 1}: Azimuth, elevation and range [rad; rad; m]

source