NRLMSISE00

The NRLMSISE00 module provides a pure-Julia implementation of the NRLMSISE00 Earth atmospheric model.

The implementation is based off of Daniel Brodo's C implementation which can be found here.

Most of the module code is for purple internal only. The density should only be retrieved through the density_nrlmsise00 function which will properly compute the required model inputs for the given time.

Unforunately because the Geomagnetic and solar flux data has no predicted component this density model is currently restricted to computing density in the past.

SatelliteDynamics.density_nrlmsise00 — Function

Computes the local atmospheric density using the NRLMSISE00 atmosphere model.

Arguments:

epc::Epoch: Epoch of computation. Used to lookup space weather data

x::AbstractArray{<:Real, 1}: Satellite state in geodetic coordinates [lon, lat, alt]

use_degrees:Bool: If true interprets geodetic inputs as being in degrees

Returns:

rho:Float64: Local atmospheric density [kg/m^3]

Notes:

Uses the gtd7d subroutine of the NRLMSISE00 to compute local atmospheric density. This subroutine includes the contribution of anomalous Oxygen in local density which is important for satellites above 500 km altitude.

References:

Picone, JM, et al. NRLMSISE-00 empirical model of the atmosphere: Statistical comparisons and scientific issues Journal of Geophysical Research: Space Physics