Space Weather

The Space Weather submodule provides classes to store and access space weather data files. In particular Solar Flux and Geomagnetic index data.

The module defines the global variables GEOMAGNETIC_DATA and SOLAR_FLUX_DATA which are loaded at runtime and

GEOMAGNETIC_DATA stores Geomagnetic Kp and Ap indicies used in models of Earth's atmosphere.

SOLAR_FLUX_DATA stores F10.7 cm solar flux data which is a measurement of solar activity and another input to most atmospheric models.

SatelliteDynamics.GeomagneticIndexData — Type

GeomagneticIndexData stores geomagnetic index data stores the geomagnetic activity Kp index and the dervied Ap indices.

Attributes

data::Dict

SatelliteDynamics.GEOMAGNETIC_DATA — Constant

Module-wide global GeomagneticIndexData. This data object is used as the default source of geomagnetic data by dynamics models if no explicit GeomagneticIndexData file is provided to those transformations.

This value can be overridden in your own code as follows:

SatelliteDynamics.SpaceWeather.GEOMAGNETIC_DATA = GeomagneticIndexData(index_file)

This global variable defaults used are pull from GFZ Potsdam FTP server: ftp://ftp.gfz-potsdam.de/pub/home/obs/kp-ap/wdc

SatelliteDynamics.KpIndices — Function

Retrieve Kp Indices for day in question

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.KpIndex — Function

Kp index for the given epoch.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.KpDailyIndex — Function

Retrieve daily Kp index for the day in question.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.ApIndices — Function

Retrieve Ap Indices for day in question

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.ApIndex — Function

Ap index for the given epoch.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.ApDailyIndex — Function

Retrieve daily Ap index for the day in question.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.SolarFluxData — Type

SolarFluxData stores solar flux index data.

The data is a dictionary keyed to the modified julian date in UT of the day of observation. All observations are taken at 20:00 UT at the of question. The values are a tuple storing the solar radio flux the F10.7cm solar radio flux. The first entry is the flux observed directly at the Ottowa radio observatory station, while the second entry is the flux value adjusted to 1 AU. The final two entries is the 81-day average, cenerted on the day entry, of the observed and adjusted solar radio flux.

Attributes

data::Dict{Int, Tuple{Float64, Float64, Float64} Stores observed F10.7cm

solar radio flux, the adjusted solar radio flux, and the 81-day average centered on the day in question.

SatelliteDynamics.SOLAR_FLUX_DATA — Constant

Module-wide global SolarFluxData object. This data object is used as the default source of Solar Flux Data by dynamics models if no explicit SolarFluxData file is provided to those transformations.

This value can be overridden in your own code as follows:

SatelliteDynamics.SpaceWeather.SOLAR_FLUX_DATA = SolarFluxData(flux_file)

This global variable defaults uses flux data provide by National Resources Canada: ftp://ftp.seismo.nrcan.gc.ca/spaceweather/solarflux/dailyflux_values/fluxtable.txt

SatelliteDynamics.f107Data — Function

F10.7 cm solar flux data on the day in question.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

Returns:

data::Tuple{Float64, Float64, Float64, Float64} Flux data on the day in question

Elements are observed, adjusted, 81-observed average, 81-adjusted average.

SatelliteDynamics.f107Observed — Function

Retrieve the F10.7 cm solar flux data observed at the day in question.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.f107Adjusted — Function

Retrieve the F10.7 cm solar flux data adjusted to 1 AU distance on the day in question.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.f107ObservedAvg — Function

Retrieve the 81-day average of the F10.7 cm solar flux data observed at the observatory centered on the day in question.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.f107AdjustedAvg — Function

Retrieve the 81-day average of the F10.7 cm solar flux data adjusted to 1 AU centered on the day in question.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1