Space Weather

The Space Weather submodule provides classes to store and access space weather data files. In particular Solar Flux and Geomagnetic index data.

The module defines the global variables GEOMAGNETIC_DATA and SOLAR_FLUX_DATA which are loaded at runtime and

GEOMAGNETIC_DATA stores Geomagnetic Kp and Ap indicies used in models of Earth's atmosphere.

SOLAR_FLUX_DATA stores F10.7 cm solar flux data which is a measurement of solar activity and another input to most atmospheric models.

SatelliteDynamics.SpaceWeatherData — Type

SpaceWeatherData stores space weather data. The data contains both geomagnetic index data and solar flux data.

Attributes

geomagnetic::SpaceWeatherData Geomagnetic index data

SatelliteDynamics.SPACE_WEATHER_DATA — Constant

Module-wide global SpaceWeatherData. This data object is used as the default source of geomagnetic data by dynamics models if no explicit SpaceWeatherData file is provided to those transformations.

This value can be overridden in your own code as follows:

SatelliteDynamics.SpaceWeather.SPACE_WEATHER_DATA = SpaceWeatherData(index_file)

This global variable defaults used are pull from the Celestrak Space Weather data repository: https://celestrak.org/SpaceData/

SatelliteDynamics.load_space_weather_data — Function

Load new space weather data into the module-wide global SpaceWeatherData structure.

Arguments:

filename::String Path to the space weather data file.

SatelliteDynamics.KpIndices — Function

Retrieve Kp Indices for day in question

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.KpIndex — Function

Kp index for the given epoch.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.KpDailyIndex — Function

Retrieve daily Kp index for the day in question.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.ApIndices — Function

Retrieve Ap Indices for day in question

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.ApIndex — Function

Ap index for the given epoch.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.ApDailyIndex — Function

Retrieve daily Ap index for the day in question.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.f107Data — Function

F10.7 cm solar flux data on the day in question.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

Returns:

data::Tuple{Float64, Float64, Float64, Float64} Flux data on the day in question

Elements are observed, adjusted, 81-observed average, 81-adjusted average.

SatelliteDynamics.f107Observed — Function

Retrieve the F10.7 cm solar flux data observed at the day in question.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.f107Adjusted — Function

Retrieve the F10.7 cm solar flux data adjusted to 1 AU distance on the day in question.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.f107ObservedAvg — Function

Retrieve the 81-day average of the F10.7 cm solar flux data observed at the observatory over the previous 81-days.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1

SatelliteDynamics.f107AdjustedAvg — Function

Retrieve the 81-day average of the F10.7 cm solar flux data adjusted to 1 AU over the previous 81-days.

Arguments:

mjd::Real Modified Julian date of desired data. Time system of input is UT1

epc::Epoch Epoch of desired input Time system of input is UT1