Integrators

The Integrators module is meant to provide built in numerical integration capabilities for the module. While the dynamics functions provided as part of of the Simulation module are compatible with the more complete set of numerical integrators provided in DifferentialEquations.jl, the integrators implemented here make other useful functionality such as integration of the Variational Equations, and incorporating impulsive maneuvers easier to implement.

Currently only the RK4 integrator is supported.

Normally it is easier to use the propagators defined in the Propagators module for simulation, but the class and method provided here can be used for the implementation of custom numerical propagators.

SatelliteDynamics.RK4 — Type

RK4 is a numerical integrator object which stores a function of signature f(epc, x,...; kwargs...) along with its associated settings parameters.

Arguments:

f::Function Function for integration.

auxParams::Dict

SatelliteDynamics.istep — Function

Performs single integration step of numerical integrator.

Arguments:

rk4::RK4 4-th order Runge-Kutta numerical integrator object

epc::Union{Real, Epoch} Absolute time of start of integration step

dt::Real Integration step size

x::AbstractArray{<:Real, 1} State vector

phi::AbstractArray{<:Real, 2} State transition matrix at start of integration step